Crate rusoto_ds

AWS Directory Service
AWS Directory Service is a web service that makes it easy for you to setup and run directories in the AWS cloud, or connect your AWS resources with an existing on-premises Microsoft Active Directory. This guide provides detailed information about AWS Directory Service operations, data types, parameters, and errors. For information about AWS Directory Services features, see AWS Directory Service and the AWS Directory Service Administration Guide.

AWS provides SDKs that consist of libraries and sample code for various programming languages and platforms (Java, Ruby, .Net, iOS, Android, etc.). The SDKs provide a convenient way to create programmatic access to AWS Directory Service and other AWS services. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

If you're using the service, you're probably looking for DirectoryServiceClient and DirectoryService.

Structs

AcceptSharedDirectoryRequest

AcceptSharedDirectoryResult

AddIpRoutesRequest

AddIpRoutesResult

AddTagsToResourceRequest

AddTagsToResourceResult

Attribute
Represents a named directory attribute.

CancelSchemaExtensionRequest

CancelSchemaExtensionResult

Computer
Contains information about a computer account in a directory.

ConditionalForwarder
Points to a remote domain with which you are setting up a trust relationship. Conditional forwarders are required in order to set up a trust relationship with another domain.

ConnectDirectoryRequest
Contains the inputs for the ConnectDirectory operation.

ConnectDirectoryResult
Contains the results of the ConnectDirectory operation.

CreateAliasRequest
Contains the inputs for the CreateAlias operation.

CreateAliasResult
Contains the results of the CreateAlias operation.

CreateComputerRequest
Contains the inputs for the CreateComputer operation.

CreateComputerResult
Contains the results for the CreateComputer operation.

CreateConditionalForwarderRequest
Initiates the creation of a conditional forwarder for your AWS Directory Service for Microsoft Active Directory. Conditional forwarders are required in order to set up a trust relationship with another domain.

CreateConditionalForwarderResult
The result of a CreateConditinalForwarder request.

CreateDirectoryRequest
Contains the inputs for the CreateDirectory operation.

CreateDirectoryResult
Contains the results of the CreateDirectory operation.

CreateLogSubscriptionRequest

CreateLogSubscriptionResult

CreateMicrosoftADRequest
Creates an AWS Managed Microsoft AD directory.

CreateMicrosoftADResult
Result of a CreateMicrosoftAD request.

CreateSnapshotRequest
Contains the inputs for the CreateSnapshot operation.

CreateSnapshotResult
Contains the results of the CreateSnapshot operation.

CreateTrustRequest
AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your AWS Managed Microsoft AD directory, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between an AWS Managed Microsoft AD directory and an external domain.

CreateTrustResult
The result of a CreateTrust request.

DeleteConditionalForwarderRequest
Deletes a conditional forwarder.

DeleteConditionalForwarderResult
The result of a DeleteConditionalForwarder request.

DeleteDirectoryRequest
Contains the inputs for the DeleteDirectory operation.

DeleteDirectoryResult
Contains the results of the DeleteDirectory operation.

DeleteLogSubscriptionRequest

DeleteLogSubscriptionResult

DeleteSnapshotRequest
Contains the inputs for the DeleteSnapshot operation.

DeleteSnapshotResult
Contains the results of the DeleteSnapshot operation.

DeleteTrustRequest
Deletes the local side of an existing trust relationship between the AWS Managed Microsoft AD directory and the external domain.

DeleteTrustResult
The result of a DeleteTrust request.

DeregisterEventTopicRequest
Removes the specified directory as a publisher to the specified SNS topic.

DeregisterEventTopicResult
The result of a DeregisterEventTopic request.

DescribeConditionalForwardersRequest
Describes a conditional forwarder.

DescribeConditionalForwardersResult
The result of a DescribeConditionalForwarder request.

DescribeDirectoriesRequest
Contains the inputs for the DescribeDirectories operation.

DescribeDirectoriesResult
Contains the results of the DescribeDirectories operation.

DescribeDomainControllersRequest

DescribeDomainControllersResult

DescribeEventTopicsRequest
Describes event topics.

DescribeEventTopicsResult
The result of a DescribeEventTopic request.

DescribeSharedDirectoriesRequest

DescribeSharedDirectoriesResult

DescribeSnapshotsRequest
Contains the inputs for the DescribeSnapshots operation.

DescribeSnapshotsResult
Contains the results of the DescribeSnapshots operation.

DescribeTrustsRequest
Describes the trust relationships for a particular AWS Managed Microsoft AD directory. If no input parameters are are provided, such as directory ID or trust ID, this request describes all the trust relationships.

DescribeTrustsResult
The result of a DescribeTrust request.

DirectoryConnectSettings
Contains information for the ConnectDirectory operation when an AD Connector directory is being created.

DirectoryConnectSettingsDescription
Contains information about an AD Connector directory.

DirectoryDescription
Contains information about an AWS Directory Service directory.

DirectoryLimits
Contains directory limit information for a region.

DirectoryServiceClient
A client for the Directory Service API.

DirectoryVpcSettings
Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

DirectoryVpcSettingsDescription
Contains information about the directory.

DisableRadiusRequest
Contains the inputs for the DisableRadius operation.

DisableRadiusResult
Contains the results of the DisableRadius operation.

DisableSsoRequest
Contains the inputs for the DisableSso operation.

DisableSsoResult
Contains the results of the DisableSso operation.

DomainController
Contains information about the domain controllers for a specified directory.

EnableRadiusRequest
Contains the inputs for the EnableRadius operation.

EnableRadiusResult
Contains the results of the EnableRadius operation.

EnableSsoRequest
Contains the inputs for the EnableSso operation.

EnableSsoResult
Contains the results of the EnableSso operation.

EventTopic
Information about SNS topic and AWS Directory Service directory associations.

GetDirectoryLimitsRequest
Contains the inputs for the GetDirectoryLimits operation.

GetDirectoryLimitsResult
Contains the results of the GetDirectoryLimits operation.

GetSnapshotLimitsRequest
Contains the inputs for the GetSnapshotLimits operation.

GetSnapshotLimitsResult
Contains the results of the GetSnapshotLimits operation.

IpRoute
IP address block. This is often the address block of the DNS server used for your on-premises domain.

IpRouteInfo
Information about one or more IP address blocks.

ListIpRoutesRequest

ListIpRoutesResult

ListLogSubscriptionsRequest

ListLogSubscriptionsResult

ListSchemaExtensionsRequest

ListSchemaExtensionsResult

ListTagsForResourceRequest

ListTagsForResourceResult

LogSubscription
Represents a log subscription, which tracks real-time data from a chosen log group to a specified destination.

OwnerDirectoryDescription
Describes the directory owner account details that have been shared to the directory consumer account.

RadiusSettings
Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

RegisterEventTopicRequest
Registers a new event topic.

RegisterEventTopicResult
The result of a RegisterEventTopic request.

RejectSharedDirectoryRequest

RejectSharedDirectoryResult

RemoveIpRoutesRequest

RemoveIpRoutesResult

RemoveTagsFromResourceRequest

RemoveTagsFromResourceResult

ResetUserPasswordRequest

ResetUserPasswordResult

RestoreFromSnapshotRequest
An object representing the inputs for the RestoreFromSnapshot operation.

RestoreFromSnapshotResult
Contains the results of the RestoreFromSnapshot operation.

SchemaExtensionInfo
Information about a schema extension.

ShareDirectoryRequest

ShareDirectoryResult

ShareTarget
Identifier that contains details about the directory consumer account.

SharedDirectory
Details about the shared directory in the directory owner account for which the share request in the directory consumer account has been accepted.

Snapshot
Describes a directory snapshot.

SnapshotLimits
Contains manual snapshot limit information for a directory.

StartSchemaExtensionRequest

StartSchemaExtensionResult

Tag
Metadata assigned to a directory consisting of a key-value pair.

Trust
Describes a trust relationship between an AWS Managed Microsoft AD directory and an external domain.

UnshareDirectoryRequest

UnshareDirectoryResult

UnshareTarget
Identifier that contains details about the directory consumer account with whom the directory is being unshared.

UpdateConditionalForwarderRequest
Updates a conditional forwarder.

UpdateConditionalForwarderResult
The result of an UpdateConditionalForwarder request.

UpdateNumberOfDomainControllersRequest

UpdateNumberOfDomainControllersResult

UpdateRadiusRequest
Contains the inputs for the UpdateRadius operation.

UpdateRadiusResult
Contains the results of the UpdateRadius operation.

UpdateTrustRequest

UpdateTrustResult

VerifyTrustRequest
Initiates the verification of an existing trust relationship between an AWS Managed Microsoft AD directory and an external domain.

VerifyTrustResult
Result of a VerifyTrust request.

Enums

AcceptSharedDirectoryError
Errors returned by AcceptSharedDirectory

AddIpRoutesError
Errors returned by AddIpRoutes

AddTagsToResourceError
Errors returned by AddTagsToResource

CancelSchemaExtensionError
Errors returned by CancelSchemaExtension

ConnectDirectoryError
Errors returned by ConnectDirectory

CreateAliasError
Errors returned by CreateAlias

CreateComputerError
Errors returned by CreateComputer

CreateConditionalForwarderError
Errors returned by CreateConditionalForwarder

CreateDirectoryError
Errors returned by CreateDirectory

CreateLogSubscriptionError
Errors returned by CreateLogSubscription

CreateMicrosoftADError
Errors returned by CreateMicrosoftAD

CreateSnapshotError
Errors returned by CreateSnapshot

CreateTrustError
Errors returned by CreateTrust

DeleteConditionalForwarderError
Errors returned by DeleteConditionalForwarder

DeleteDirectoryError
Errors returned by DeleteDirectory

DeleteLogSubscriptionError
Errors returned by DeleteLogSubscription

DeleteSnapshotError
Errors returned by DeleteSnapshot

DeleteTrustError
Errors returned by DeleteTrust

DeregisterEventTopicError
Errors returned by DeregisterEventTopic

DescribeConditionalForwardersError
Errors returned by DescribeConditionalForwarders

DescribeDirectoriesError
Errors returned by DescribeDirectories

DescribeDomainControllersError
Errors returned by DescribeDomainControllers

DescribeEventTopicsError
Errors returned by DescribeEventTopics

DescribeSharedDirectoriesError
Errors returned by DescribeSharedDirectories

DescribeSnapshotsError
Errors returned by DescribeSnapshots

DescribeTrustsError
Errors returned by DescribeTrusts

DisableRadiusError
Errors returned by DisableRadius

DisableSsoError
Errors returned by DisableSso

EnableRadiusError
Errors returned by EnableRadius

EnableSsoError
Errors returned by EnableSso

GetDirectoryLimitsError
Errors returned by GetDirectoryLimits

GetSnapshotLimitsError
Errors returned by GetSnapshotLimits

ListIpRoutesError
Errors returned by ListIpRoutes

ListLogSubscriptionsError
Errors returned by ListLogSubscriptions

ListSchemaExtensionsError
Errors returned by ListSchemaExtensions

ListTagsForResourceError
Errors returned by ListTagsForResource

RegisterEventTopicError
Errors returned by RegisterEventTopic

RejectSharedDirectoryError
Errors returned by RejectSharedDirectory

RemoveIpRoutesError
Errors returned by RemoveIpRoutes

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetUserPasswordError
Errors returned by ResetUserPassword

RestoreFromSnapshotError
Errors returned by RestoreFromSnapshot

ShareDirectoryError
Errors returned by ShareDirectory

StartSchemaExtensionError
Errors returned by StartSchemaExtension

UnshareDirectoryError
Errors returned by UnshareDirectory

UpdateConditionalForwarderError
Errors returned by UpdateConditionalForwarder

UpdateNumberOfDomainControllersError
Errors returned by UpdateNumberOfDomainControllers

UpdateRadiusError
Errors returned by UpdateRadius

UpdateTrustError
Errors returned by UpdateTrust

VerifyTrustError
Errors returned by VerifyTrust

Traits

DirectoryService
Trait representing the capabilities of the Directory Service API. Directory Service clients implement this trait.