Struct rusoto_ds::RadiusSettings

pub struct RadiusSettings {
    pub authentication_protocol: Option<String>,
    pub display_label: Option<String>,
    pub radius_port: Option<i64>,
    pub radius_retries: Option<i64>,
    pub radius_servers: Option<Vec<String>>,
    pub radius_timeout: Option<i64>,
    pub shared_secret: Option<String>,
    pub use_same_username: Option<bool>,
}

Contains information about a Remote Authentication Dial In User Service (RADIUS) server.

Fields
authentication_protocol: Option<String>

The protocol specified for your RADIUS endpoints.

display_label: Option<String>

Not currently used.

radius_port: Option<i64>

The port that your RADIUS server is using for communications. Your on-premises network must allow inbound traffic over this port from the AWS Directory Service servers.

radius_retries: Option<i64>

The maximum number of times that communication with the RADIUS server is attempted.

radius_servers: Option<Vec<String>>

An array of strings that contains the IP addresses of the RADIUS server endpoints, or the IP addresses of your RADIUS server load balancer.

radius_timeout: Option<i64>

The amount of time, in seconds, to wait for the RADIUS server to respond.

shared_secret: Option<String>

Not currently used.

use_same_username: Option<bool>

Not currently used.