Trait rusoto_ds::DirectoryService

pub trait DirectoryService {
    fn accept_shared_directory(
        &self, 
        input: AcceptSharedDirectoryRequest
    ) -> RusotoFuture<AcceptSharedDirectoryResult, AcceptSharedDirectoryError>;

    fn add_ip_routes(
        &self, 
        input: AddIpRoutesRequest
    ) -> RusotoFuture<AddIpRoutesResult, AddIpRoutesError>;

    fn add_tags_to_resource(
        &self, 
        input: AddTagsToResourceRequest
    ) -> RusotoFuture<AddTagsToResourceResult, AddTagsToResourceError>;

    fn cancel_schema_extension(
        &self, 
        input: CancelSchemaExtensionRequest
    ) -> RusotoFuture<CancelSchemaExtensionResult, CancelSchemaExtensionError>;

    fn connect_directory(
        &self, 
        input: ConnectDirectoryRequest
    ) -> RusotoFuture<ConnectDirectoryResult, ConnectDirectoryError>;

    fn create_alias(
        &self, 
        input: CreateAliasRequest
    ) -> RusotoFuture<CreateAliasResult, CreateAliasError>;

    fn create_computer(
        &self, 
        input: CreateComputerRequest
    ) -> RusotoFuture<CreateComputerResult, CreateComputerError>;

    fn create_conditional_forwarder(
        &self, 
        input: CreateConditionalForwarderRequest
    ) -> RusotoFuture<CreateConditionalForwarderResult, CreateConditionalForwarderError>;

    fn create_directory(
        &self, 
        input: CreateDirectoryRequest
    ) -> RusotoFuture<CreateDirectoryResult, CreateDirectoryError>;

    fn create_log_subscription(
        &self, 
        input: CreateLogSubscriptionRequest
    ) -> RusotoFuture<CreateLogSubscriptionResult, CreateLogSubscriptionError>;

    fn create_microsoft_ad(
        &self, 
        input: CreateMicrosoftADRequest
    ) -> RusotoFuture<CreateMicrosoftADResult, CreateMicrosoftADError>;

    fn create_snapshot(
        &self, 
        input: CreateSnapshotRequest
    ) -> RusotoFuture<CreateSnapshotResult, CreateSnapshotError>;

    fn create_trust(
        &self, 
        input: CreateTrustRequest
    ) -> RusotoFuture<CreateTrustResult, CreateTrustError>;

    fn delete_conditional_forwarder(
        &self, 
        input: DeleteConditionalForwarderRequest
    ) -> RusotoFuture<DeleteConditionalForwarderResult, DeleteConditionalForwarderError>;

    fn delete_directory(
        &self, 
        input: DeleteDirectoryRequest
    ) -> RusotoFuture<DeleteDirectoryResult, DeleteDirectoryError>;

    fn delete_log_subscription(
        &self, 
        input: DeleteLogSubscriptionRequest
    ) -> RusotoFuture<DeleteLogSubscriptionResult, DeleteLogSubscriptionError>;

    fn delete_snapshot(
        &self, 
        input: DeleteSnapshotRequest
    ) -> RusotoFuture<DeleteSnapshotResult, DeleteSnapshotError>;

    fn delete_trust(
        &self, 
        input: DeleteTrustRequest
    ) -> RusotoFuture<DeleteTrustResult, DeleteTrustError>;

    fn deregister_event_topic(
        &self, 
        input: DeregisterEventTopicRequest
    ) -> RusotoFuture<DeregisterEventTopicResult, DeregisterEventTopicError>;

    fn describe_conditional_forwarders(
        &self, 
        input: DescribeConditionalForwardersRequest
    ) -> RusotoFuture<DescribeConditionalForwardersResult, DescribeConditionalForwardersError>;

    fn describe_directories(
        &self, 
        input: DescribeDirectoriesRequest
    ) -> RusotoFuture<DescribeDirectoriesResult, DescribeDirectoriesError>;

    fn describe_domain_controllers(
        &self, 
        input: DescribeDomainControllersRequest
    ) -> RusotoFuture<DescribeDomainControllersResult, DescribeDomainControllersError>;

    fn describe_event_topics(
        &self, 
        input: DescribeEventTopicsRequest
    ) -> RusotoFuture<DescribeEventTopicsResult, DescribeEventTopicsError>;

    fn describe_shared_directories(
        &self, 
        input: DescribeSharedDirectoriesRequest
    ) -> RusotoFuture<DescribeSharedDirectoriesResult, DescribeSharedDirectoriesError>;

    fn describe_snapshots(
        &self, 
        input: DescribeSnapshotsRequest
    ) -> RusotoFuture<DescribeSnapshotsResult, DescribeSnapshotsError>;

    fn describe_trusts(
        &self, 
        input: DescribeTrustsRequest
    ) -> RusotoFuture<DescribeTrustsResult, DescribeTrustsError>;

    fn disable_radius(
        &self, 
        input: DisableRadiusRequest
    ) -> RusotoFuture<DisableRadiusResult, DisableRadiusError>;

    fn disable_sso(
        &self, 
        input: DisableSsoRequest
    ) -> RusotoFuture<DisableSsoResult, DisableSsoError>;

    fn enable_radius(
        &self, 
        input: EnableRadiusRequest
    ) -> RusotoFuture<EnableRadiusResult, EnableRadiusError>;

    fn enable_sso(
        &self, 
        input: EnableSsoRequest
    ) -> RusotoFuture<EnableSsoResult, EnableSsoError>;

    fn get_directory_limits(
        &self
    ) -> RusotoFuture<GetDirectoryLimitsResult, GetDirectoryLimitsError>;

    fn get_snapshot_limits(
        &self, 
        input: GetSnapshotLimitsRequest
    ) -> RusotoFuture<GetSnapshotLimitsResult, GetSnapshotLimitsError>;

    fn list_ip_routes(
        &self, 
        input: ListIpRoutesRequest
    ) -> RusotoFuture<ListIpRoutesResult, ListIpRoutesError>;

    fn list_log_subscriptions(
        &self, 
        input: ListLogSubscriptionsRequest
    ) -> RusotoFuture<ListLogSubscriptionsResult, ListLogSubscriptionsError>;

    fn list_schema_extensions(
        &self, 
        input: ListSchemaExtensionsRequest
    ) -> RusotoFuture<ListSchemaExtensionsResult, ListSchemaExtensionsError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResult, ListTagsForResourceError>;

    fn register_event_topic(
        &self, 
        input: RegisterEventTopicRequest
    ) -> RusotoFuture<RegisterEventTopicResult, RegisterEventTopicError>;

    fn reject_shared_directory(
        &self, 
        input: RejectSharedDirectoryRequest
    ) -> RusotoFuture<RejectSharedDirectoryResult, RejectSharedDirectoryError>;

    fn remove_ip_routes(
        &self, 
        input: RemoveIpRoutesRequest
    ) -> RusotoFuture<RemoveIpRoutesResult, RemoveIpRoutesError>;

    fn remove_tags_from_resource(
        &self, 
        input: RemoveTagsFromResourceRequest
    ) -> RusotoFuture<RemoveTagsFromResourceResult, RemoveTagsFromResourceError>;

    fn reset_user_password(
        &self, 
        input: ResetUserPasswordRequest
    ) -> RusotoFuture<ResetUserPasswordResult, ResetUserPasswordError>;

    fn restore_from_snapshot(
        &self, 
        input: RestoreFromSnapshotRequest
    ) -> RusotoFuture<RestoreFromSnapshotResult, RestoreFromSnapshotError>;

    fn share_directory(
        &self, 
        input: ShareDirectoryRequest
    ) -> RusotoFuture<ShareDirectoryResult, ShareDirectoryError>;

    fn start_schema_extension(
        &self, 
        input: StartSchemaExtensionRequest
    ) -> RusotoFuture<StartSchemaExtensionResult, StartSchemaExtensionError>;

    fn unshare_directory(
        &self, 
        input: UnshareDirectoryRequest
    ) -> RusotoFuture<UnshareDirectoryResult, UnshareDirectoryError>;

    fn update_conditional_forwarder(
        &self, 
        input: UpdateConditionalForwarderRequest
    ) -> RusotoFuture<UpdateConditionalForwarderResult, UpdateConditionalForwarderError>;

    fn update_number_of_domain_controllers(
        &self, 
        input: UpdateNumberOfDomainControllersRequest
    ) -> RusotoFuture<UpdateNumberOfDomainControllersResult, UpdateNumberOfDomainControllersError>;

    fn update_radius(
        &self, 
        input: UpdateRadiusRequest
    ) -> RusotoFuture<UpdateRadiusResult, UpdateRadiusError>;

    fn update_trust(
        &self, 
        input: UpdateTrustRequest
    ) -> RusotoFuture<UpdateTrustResult, UpdateTrustError>;

    fn verify_trust(
        &self, 
        input: VerifyTrustRequest
    ) -> RusotoFuture<VerifyTrustResult, VerifyTrustError>;
}

Trait representing the capabilities of the Directory Service API. Directory Service clients implement this trait.

Required methods

fn accept_shared_directory(
&self,
input: AcceptSharedDirectoryRequest
) -> RusotoFuture<AcceptSharedDirectoryResult, AcceptSharedDirectoryError>

Accepts a directory sharing request that was sent from the directory owner account.

fn add_ip_routes(
&self,
input: AddIpRoutesRequest
) -> RusotoFuture<AddIpRoutesResult, AddIpRoutesError>

If the DNS server for your on-premises domain uses a publicly addressable IP address, you must add a CIDR address block to correctly route traffic to and from your Microsoft AD on Amazon Web Services. AddIpRoutes adds this address block. You can also use AddIpRoutes to facilitate routing traffic that uses public IP ranges from your Microsoft AD on AWS to a peer VPC.

Before you call AddIpRoutes, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the AddIpRoutes operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

fn add_tags_to_resource(
&self,
input: AddTagsToResourceRequest
) -> RusotoFuture<AddTagsToResourceResult, AddTagsToResourceError>

Adds or overwrites one or more tags for the specified directory. Each directory can have a maximum of 50 tags. Each tag consists of a key and optional value. Tag keys must be unique to each resource.

fn cancel_schema_extension(
&self,
input: CancelSchemaExtensionRequest
) -> RusotoFuture<CancelSchemaExtensionResult, CancelSchemaExtensionError>

Cancels an in-progress schema extension to a Microsoft AD directory. Once a schema extension has started replicating to all domain controllers, the task can no longer be canceled. A schema extension can be canceled during any of the following states; Initializing, CreatingSnapshot, and UpdatingSchema.

fn connect_directory(
&self,
input: ConnectDirectoryRequest
) -> RusotoFuture<ConnectDirectoryResult, ConnectDirectoryError>

Creates an AD Connector to connect to an on-premises directory.

Before you call ConnectDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the ConnectDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

fn create_alias(
&self,
input: CreateAliasRequest
) -> RusotoFuture<CreateAliasResult, CreateAliasError>

Creates an alias for a directory and assigns the alias to the directory. The alias is used to construct the access URL for the directory, such as http://<alias>.awsapps.com.

After an alias has been created, it cannot be deleted or reused, so this operation should only be used when absolutely necessary.

fn create_computer(
&self,
input: CreateComputerRequest
) -> RusotoFuture<CreateComputerResult, CreateComputerError>

Creates a computer account in the specified directory, and joins the computer to the directory.

fn create_conditional_forwarder(
&self,
input: CreateConditionalForwarderRequest
) -> RusotoFuture<CreateConditionalForwarderResult, CreateConditionalForwarderError>

Creates a conditional forwarder associated with your AWS directory. Conditional forwarders are required in order to set up a trust relationship with another domain. The conditional forwarder points to the trusted domain.

fn create_directory(
&self,
input: CreateDirectoryRequest
) -> RusotoFuture<CreateDirectoryResult, CreateDirectoryError>

Creates a Simple AD directory.

Before you call CreateDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

fn create_log_subscription(
&self,
input: CreateLogSubscriptionRequest
) -> RusotoFuture<CreateLogSubscriptionResult, CreateLogSubscriptionError>

Creates a subscription to forward real time Directory Service domain controller security logs to the specified CloudWatch log group in your AWS account.

fn create_microsoft_ad(
&self,
input: CreateMicrosoftADRequest
) -> RusotoFuture<CreateMicrosoftADResult, CreateMicrosoftADError>

Creates an AWS Managed Microsoft AD directory.

Before you call CreateMicrosoftAD, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateMicrosoftAD operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

fn create_snapshot(
&self,
input: CreateSnapshotRequest
) -> RusotoFuture<CreateSnapshotResult, CreateSnapshotError>

Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS cloud.

You cannot take snapshots of AD Connector directories.

fn create_trust(
&self,
input: CreateTrustRequest
) -> RusotoFuture<CreateTrustResult, CreateTrustError>

AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your AWS Managed Microsoft AD directory, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between an AWS Managed Microsoft AD directory and an external domain. You can create either a forest trust or an external trust.

fn delete_conditional_forwarder(
&self,
input: DeleteConditionalForwarderRequest
) -> RusotoFuture<DeleteConditionalForwarderResult, DeleteConditionalForwarderError>

Deletes a conditional forwarder that has been set up for your AWS directory.

fn delete_directory(
&self,
input: DeleteDirectoryRequest
) -> RusotoFuture<DeleteDirectoryResult, DeleteDirectoryError>

Deletes an AWS Directory Service directory.

Before you call DeleteDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the DeleteDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

fn delete_log_subscription(
&self,
input: DeleteLogSubscriptionRequest
) -> RusotoFuture<DeleteLogSubscriptionResult, DeleteLogSubscriptionError>

Deletes the specified log subscription.

fn delete_snapshot(
&self,
input: DeleteSnapshotRequest
) -> RusotoFuture<DeleteSnapshotResult, DeleteSnapshotError>

Deletes a directory snapshot.

fn delete_trust(
&self,
input: DeleteTrustRequest
) -> RusotoFuture<DeleteTrustResult, DeleteTrustError>

Deletes an existing trust relationship between your AWS Managed Microsoft AD directory and an external domain.

fn deregister_event_topic(
&self,
input: DeregisterEventTopicRequest
) -> RusotoFuture<DeregisterEventTopicResult, DeregisterEventTopicError>

Removes the specified directory as a publisher to the specified SNS topic.

fn describe_conditional_forwarders(
&self,
input: DescribeConditionalForwardersRequest
) -> RusotoFuture<DescribeConditionalForwardersResult, DescribeConditionalForwardersError>

Obtains information about the conditional forwarders for this account.

If no input parameters are provided for RemoteDomainNames, this request describes all conditional forwarders for the specified directory ID.

fn describe_directories(
&self,
input: DescribeDirectoriesRequest
) -> RusotoFuture<DescribeDirectoriesResult, DescribeDirectoriesError>

Obtains information about the directories that belong to this account.

You can retrieve information about specific directories by passing the directory identifiers in the DirectoryIds parameter. Otherwise, all directories that belong to the current account are returned.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeDirectoriesResult.NextToken member contains a token that you pass in the next call to DescribeDirectories to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

fn describe_domain_controllers(
&self,
input: DescribeDomainControllersRequest
) -> RusotoFuture<DescribeDomainControllersResult, DescribeDomainControllersError>

Provides information about any domain controllers in your directory.

fn describe_event_topics(
&self,
input: DescribeEventTopicsRequest
) -> RusotoFuture<DescribeEventTopicsResult, DescribeEventTopicsError>

Obtains information about which SNS topics receive status messages from the specified directory.

If no input parameters are provided, such as DirectoryId or TopicName, this request describes all of the associations in the account.

fn describe_shared_directories(
&self,
input: DescribeSharedDirectoriesRequest
) -> RusotoFuture<DescribeSharedDirectoriesResult, DescribeSharedDirectoriesError>

Returns the shared directories in your account.

fn describe_snapshots(
&self,
input: DescribeSnapshotsRequest
) -> RusotoFuture<DescribeSnapshotsResult, DescribeSnapshotsError>

Obtains information about the directory snapshots that belong to this account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeSnapshots.NextToken member contains a token that you pass in the next call to DescribeSnapshots to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

fn describe_trusts(
&self,
input: DescribeTrustsRequest
) -> RusotoFuture<DescribeTrustsResult, DescribeTrustsError>

Obtains information about the trust relationships for this account.

If no input parameters are provided, such as DirectoryId or TrustIds, this request describes all the trust relationships belonging to the account.

fn disable_radius(
&self,
input: DisableRadiusRequest
) -> RusotoFuture<DisableRadiusResult, DisableRadiusError>

Disables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory.

fn disable_sso(
&self,
input: DisableSsoRequest
) -> RusotoFuture<DisableSsoResult, DisableSsoError>

Disables single-sign on for a directory.

fn enable_radius(
&self,
input: EnableRadiusRequest
) -> RusotoFuture<EnableRadiusResult, EnableRadiusError>

Enables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory.

fn enable_sso(
&self,
input: EnableSsoRequest
) -> RusotoFuture<EnableSsoResult, EnableSsoError>

Enables single sign-on for a directory.

fn get_directory_limits(
&self
) -> RusotoFuture<GetDirectoryLimitsResult, GetDirectoryLimitsError>

Obtains directory limit information for the current region.

fn get_snapshot_limits(
&self,
input: GetSnapshotLimitsRequest
) -> RusotoFuture<GetSnapshotLimitsResult, GetSnapshotLimitsError>

Obtains the manual snapshot limits for a directory.

fn list_ip_routes(
&self,
input: ListIpRoutesRequest
) -> RusotoFuture<ListIpRoutesResult, ListIpRoutesError>

Lists the address blocks that you have added to a directory.

fn list_log_subscriptions(
&self,
input: ListLogSubscriptionsRequest
) -> RusotoFuture<ListLogSubscriptionsResult, ListLogSubscriptionsError>

Lists the active log subscriptions for the AWS account.

fn list_schema_extensions(
&self,
input: ListSchemaExtensionsRequest
) -> RusotoFuture<ListSchemaExtensionsResult, ListSchemaExtensionsError>

Lists all schema extensions applied to a Microsoft AD Directory.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResult, ListTagsForResourceError>

Lists all tags on a directory.

fn register_event_topic(
&self,
input: RegisterEventTopicRequest
) -> RusotoFuture<RegisterEventTopicResult, RegisterEventTopicError>

Associates a directory with an SNS topic. This establishes the directory as a publisher to the specified SNS topic. You can then receive email or text (SMS) messages when the status of your directory changes. You get notified if your directory goes from an Active status to an Impaired or Inoperable status. You also receive a notification when the directory returns to an Active status.

fn reject_shared_directory(
&self,
input: RejectSharedDirectoryRequest
) -> RusotoFuture<RejectSharedDirectoryResult, RejectSharedDirectoryError>

Rejects a directory sharing request that was sent from the directory owner account.

fn remove_ip_routes(
&self,
input: RemoveIpRoutesRequest
) -> RusotoFuture<RemoveIpRoutesResult, RemoveIpRoutesError>

Removes IP address blocks from a directory.

fn remove_tags_from_resource(
&self,
input: RemoveTagsFromResourceRequest
) -> RusotoFuture<RemoveTagsFromResourceResult, RemoveTagsFromResourceError>

Removes tags from a directory.

fn reset_user_password(
&self,
input: ResetUserPasswordRequest
) -> RusotoFuture<ResetUserPasswordResult, ResetUserPasswordError>

Resets the password for any user in your AWS Managed Microsoft AD or Simple AD directory.

fn restore_from_snapshot(
&self,
input: RestoreFromSnapshotRequest
) -> RusotoFuture<RestoreFromSnapshotResult, RestoreFromSnapshotError>

Restores a directory using an existing directory snapshot.

When you restore a directory from a snapshot, any changes made to the directory after the snapshot date are overwritten.

This action returns as soon as the restore operation is initiated. You can monitor the progress of the restore operation by calling the DescribeDirectories operation with the directory identifier. When the DirectoryDescription.Stage value changes to Active, the restore operation is complete.

fn share_directory(
&self,
input: ShareDirectoryRequest
) -> RusotoFuture<ShareDirectoryResult, ShareDirectoryError>

Shares a specified directory (DirectoryId) in your AWS account (directory owner) with another AWS account (directory consumer). With this operation you can use your directory from any AWS account and from any Amazon VPC within an AWS Region.

When you share your AWS Managed Microsoft AD directory, AWS Directory Service creates a shared directory in the directory consumer account. This shared directory contains the metadata to provide access to the directory within the directory owner account. The shared directory is visible in all VPCs in the directory consumer account.

The ShareMethod parameter determines whether the specified directory can be shared between AWS accounts inside the same AWS organization (ORGANIZATIONS). It also determines whether you can share the directory with any other AWS account either inside or outside of the organization (HANDSHAKE).

The ShareNotes parameter is only used when HANDSHAKE is called, which sends a directory sharing request to the directory consumer.

fn start_schema_extension(
&self,
input: StartSchemaExtensionRequest
) -> RusotoFuture<StartSchemaExtensionResult, StartSchemaExtensionError>

Applies a schema extension to a Microsoft AD directory.

fn unshare_directory(
&self,
input: UnshareDirectoryRequest
) -> RusotoFuture<UnshareDirectoryResult, UnshareDirectoryError>

Stops the directory sharing between the directory owner and consumer accounts.

fn update_conditional_forwarder(
&self,
input: UpdateConditionalForwarderRequest
) -> RusotoFuture<UpdateConditionalForwarderResult, UpdateConditionalForwarderError>

Updates a conditional forwarder that has been set up for your AWS directory.

fn update_number_of_domain_controllers(
&self,
input: UpdateNumberOfDomainControllersRequest
) -> RusotoFuture<UpdateNumberOfDomainControllersResult, UpdateNumberOfDomainControllersError>

Adds or removes domain controllers to or from the directory. Based on the difference between current value and new value (provided through this API call), domain controllers will be added or removed. It may take up to 45 minutes for any new domain controllers to become fully active once the requested number of domain controllers is updated. During this time, you cannot make another update request.

fn update_radius(
&self,
input: UpdateRadiusRequest
) -> RusotoFuture<UpdateRadiusResult, UpdateRadiusError>

Updates the Remote Authentication Dial In User Service (RADIUS) server information for an AD Connector or Microsoft AD directory.

fn update_trust(
&self,
input: UpdateTrustRequest
) -> RusotoFuture<UpdateTrustResult, UpdateTrustError>

Updates the trust that has been set up between your AWS Managed Microsoft AD directory and an on-premises Active Directory.

fn verify_trust(
&self,
input: VerifyTrustRequest
) -> RusotoFuture<VerifyTrustResult, VerifyTrustError>

AWS Directory Service for Microsoft Active Directory allows you to configure and verify trust relationships.

This action verifies a trust relationship between your AWS Managed Microsoft AD directory and an external domain.

Loading content...