Struct rusoto_ds::DirectoryServiceClient

pub struct DirectoryServiceClient { /* fields omitted */ }

A client for the Directory Service API.

Methods

impl DirectoryServiceClient

pub fn new(region: Region) -> DirectoryServiceClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DirectoryServiceClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,